(PA-3755) use rsync3 on mac OS >= 11
mac OS >= 11 comes with rsync 3 while older versions
have rsyn 2.6.9 installed. To ensure that the signature
of the files is maintained, we need to preserve
extended attributes. On rsync 2 we need to use the
`--extended-attributes` flag while on rsync 3 we
need to use `--xattrs` and specify to the remote host
to use rsync 3 using `--rsync-path`